Se cambia el tipo de dato para solucionar el error , EF Core intenta mapear el valor FLOAT de Discount a double

+++ Sofka.Piguinera.Cotizacion/Sofka.Piguinera.Cotizacion/Models/Persistence/BookPersistence.cs
@@ -10,7 +10,7 @@
         public int? Quantity { get; set; }
         public byte Type { get; set; }
         public double UnitPrice { get; set; }
-        public float Discount { get; set; }
+        public double Discount { get; set; } // Cambiado de float a double
 
     }
 }
